+++ src/main/java/hibernate/CreateQueryBuilder.java
@@ -1,12 +1,12 @@
 package hibernate;
 
 import hibernate.entity.EntityClass;
-import hibernate.entity.EntityColumn;
-import hibernate.entity.EntityColumnFactory;
-import hibernate.entity.strategy.ColumnOptionGenerateStrategy;
-import hibernate.entity.strategy.IdIdentityOptionGenerateStrategy;
-import hibernate.entity.strategy.NotNullOptionGenerateStrategy;
-import hibernate.entity.strategy.PrimaryKetOptionGenerateStrategy;
+import hibernate.entity.column.EntityColumn;
+import hibernate.entity.column.EntityColumnFactory;
+import hibernate.strategy.ColumnOptionGenerateStrategy;
+import hibernate.strategy.IdIdentityOptionGenerateStrategy;
+import hibernate.strategy.NotNullOptionGenerateStrategy;
+import hibernate.strategy.PrimaryKetOptionGenerateStrategy;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
