+++ src/main/java/hibernate/entity/column/EntityColumn.java
@@ -1,4 +1,4 @@
-package hibernate.entity;
+package hibernate.entity.column;
 
 import jakarta.persistence.GenerationType;
