+++ src/main/java/hibernate/strategy/IdIdentityOptionGenerateStrategy.java
@@ -1,6 +1,6 @@
-package hibernate.entity.strategy;
+package hibernate.strategy;
 
-import hibernate.entity.EntityColumn;
+import hibernate.entity.column.EntityColumn;
 import jakarta.persistence.GenerationType;
 
 public class IdIdentityOptionGenerateStrategy implements ColumnOptionGenerateStrategy {
