+++ src/test/java/hibernate/entity/column/ColumnTypeTest.java
@@ -1,5 +1,6 @@
-package hibernate.entity;
+package hibernate.entity.column;
 
+import hibernate.entity.column.ColumnType;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
