+++ src/test/java/hibernate/entity/column/EntityColumnFactoryTest.java
@@ -1,5 +1,9 @@
-package hibernate.entity;
+package hibernate.entity.column;
 
+import hibernate.entity.column.EntityColumn;
+import hibernate.entity.column.EntityColumnFactory;
+import hibernate.entity.column.EntityField;
+import hibernate.entity.column.EntityId;
 import jakarta.persistence.Id;
 import jakarta.persistence.Transient;
 import org.junit.jupiter.api.Test;
