+++ src/test/java/hibernate/entity/column/EntityFieldTest.java
@@ -1,7 +1,9 @@
-package hibernate.entity;
+package hibernate.entity.column;
 
 import domain.Person;
 import domain.Person2;
+import hibernate.entity.column.ColumnType;
+import hibernate.entity.column.EntityField;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
